backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
openai==0.28.0
python-jose==3.3.0
pydantic==1.10.13
python-multipart==0.0.9 